Embedded notes 2.0 (1/2 commits)
Embedded notes have a border and title of the note's name. Since Obsidian by default shows the *entire* file, a feature was implemented to give embeds a scrolling function if it is too long (over 640 px). This works in reading and editing/preview mode.

The border of embedded notes was dubbled to provide more visual feedback that it's an embed.